Move FCNNTrainer class to FCNN.py
Trying to keep the train.py file de-cluttered, and it makes more
sense to have the Trainer classes in the same file as their respective
model classes.
